Changed embed params to t and d, fixed signup system, added upload rate limiter, added system to delete images which were not visited for longer than 15 days'

+++ Uploader/lapis/apps/admin.lua
@@ -9,6 +9,7 @@ local bcrypt = require("bcrypt")
 local validate = require("lapis.validate")
 local json_params = require("lapis.application").json_params
 local util = require("lapis.util")
+local http = require("lapis.nginx.http")
 local respond_to = require("lapis.application").respond_to
 local rounds = 10
 
@@ -18,7 +19,7 @@ validate.validate_functions.low_or_eq = function(input, num)
     local a = tonumber(input)
     print(type(a))
     if  a > num then
-        return input.." must be smaller or equal to "..num
+        return false, input.." must be smaller or equal to "..num
     elseif a <= num then
         return true
     end
@@ -26,7 +27,7 @@ end
 validate.validate_functions.big_or_eq = function(input, num)
     local a = tonumber(input)
     if a < num then
-        return input.." must be bigger or equal to "..num
+        return false, input.." must be bigger or equal to "..num
     elseif a >= num then
         return true
     end
@@ -48,7 +49,7 @@ local function checkValid(sess)
 end
 local function checkAdmin(sess)
     local sel = db.select("perms FROM users WHERE username IN ( SELECT sessfor FROM sessions WHERE sessid = ? )", sess)
-    if sel[1] and sel[1].perms == "admin" then
+    if sel[1] and sel[1].perms >= 3 then
         return true
     else
         return false
@@ -83,7 +84,7 @@ app:match("/admin/login", respond_to({
     POST = json_params(function(self)
         if self.params.username and self.params.passwd then
             local dbr = db.select("* FROM USERS WHERE username = ?", self.params.username)
-            if dbr[1] ~= nil and bcrypt.verify(self.params.passwd, dbr[1].passwd) then
+            if dbr[1] ~= nil and bcrypt.verify(self.params.passwd, dbr[1].passwd) and dbr[1].perms >= 3 then
                 db.delete("sessions", "sessfor = ?", self.params.username)
                 local sess = string.random(20)
                 db.insert("sessions", {
@@ -114,20 +115,30 @@ app:match("/admin/createinvite", respond_to({
         })
         if checkValid(self.session.USER) and checkAdmin(self.session.USER)  then
             local invt = {}
-            for i = 1, self.params.amount do
-                local ainvite = "INVT-"..string.random(20)
-                local sel = db.select("invite FROM invites WHERE invite = ?", ainvite)
-                if not sel[1] then
-                    db.insert("invites", {
-                        invite = ainvite
-                    })
-                    table.insert(invt, ainvite)
-                    ngx.sleep(1)
-                else
-                    i = i - 1
+            --local ainvite = "INVT-"..string.random(20)
+            
+            local req = http.simple("https://www.random.org/strings/?num="..self.params.amount.."&len=20&digits=on&upperalpha=on&loweralpha=on&unique=on&format=plain&rnd=new")
+            local invt = string.split(req, "\n")
+            local FINVT = {}
+            for i,v in pairs(invt) do
+                v = "INVT-"..v
+                local sel = db.select("invite FROM invites WHERE invite = ?", v)
+                if sel[1] then -- cringe
+                    v = "INVT-"..http.simple("https://www.random.org/strings/?num=1&len=20&digits=on&upperalpha=on&loweralpha=on&unique=on&format=plain&rnd=new")
                 end
+                db.insert("invites", {
+                    invite = v
+                })
+                table.insert(FINVT, v)
             end
-            return { layout = false, json = { msg = invt}}
+            --[[if not sel[1] then
+                    
+                    
+                    table.insert(invt, ainvite)
+                    --ngx.sleep(1)
+                
+            end]]
+            return { layout = false, json = { msg = FINVT }}
         end
         
     end),
@@ -140,9 +151,10 @@ app:match("/admin/deleteimg", respond_to({
             if res[1] then
                 local adate = os.date("%A %x %X", res[1].upstamp)
                 local uploader = res[1].uploader
-                print(adate)
-                assert(os.remove("/srv/lapis/html/files/"..self.params.imgurl))
-                local res = db.delete("images", "imgurl = ?", self.params.imgurl)
+                assert(os.remove('/srv/lapis/html/files/'..self.params.imgurl))
+                db.delete('images', {
+                    imgurl = self.params.imgurl
+                })
                 return { status = 200, json = { resp = "File "..self.params.imgurl.." has been successfully deleted.", meta = {author = "Uploaded by: "..uploader, date = "Uploaded On: "..adate}}}
             else return {status = 404}
             end
@@ -155,6 +167,15 @@ app:match("/admin/deluser", respond_to({
     DELETE = json_params(function(self)
         if checkValid(self.session.USER) and self.params.username and checkAdmin(self.session.USER) then
             local res = db.delete("users", "username = ?", self.params.username)
+            local sel = db.select("imgurl FROM images WHERE uploader = ?", self.params.username)
+            if sel[1] then
+                for i,v in pairs(sel) do
+                    os.remove("/srv/lapis/html/files/"..v.imgurl)
+                end
+                db.delete('images', {
+                    uploader = self.params.username 
+                })
+            end
             return { status = 200, json = {aff = res.affected_rows}}
         else
             return { status = 401, redirect_to = "/admin/login"}
